SpeechMaticsUsage/requirement.txt
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
docopt==0.6.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
jiwer==3.0.5
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.5.0
numpy==2.1.3
pandas==2.2.3
polling2==0.5.0
pyannote.core==5.0.0
pyannote.database==5.1.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.11.6
rich==13.9.4
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
speechmatics-python==3.0.0
tabulate==0.9.0
tenacity==8.2.3
toml==0.10.2
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
websockets==14.1
